Move to logger module
