+++ qcodes/__init__.py
@@ -6,30 +6,13 @@
 
 from qcodes.config import Config
 from qcodes.logger import start_all_logging
+from qcodes.logger.logger import _conditionally_start_all_logging
 from qcodes.utils.helpers import add_to_spyder_UMR_excludelist
 from .version import __version__
 
 config: Config = Config()
 
 # start logging if work_station is configured or
-def _conditionally_start_all_logging():
-    def start_logging_on_import() -> bool:
-        return (
-            config.GUID_components.location != 0 and
-            config.GUID_components.work_station != 0 and
-            config.telemetry.instrumentation_key != \
-                "00000000-0000-0000-0000-000000000000"
-            or config.logger.start_logging_on_import
-        )
-
-    def running_in_test_or_tool() -> bool:
-        import sys
-        tools = (
-            'pytest.py', 'pytest', '_jb_pytest_runner.py', 'testlauncher.py' )
-        return any(sys.argv[0].endswith(tool) for tool in tools)
-
-    if start_logging_on_import() and not running_in_test_or_tool():
-        start_all_logging()
 
 _conditionally_start_all_logging()
